Add evolution module. Move update package from application to evolution. Move migration package from application to evolution. Move creation and instantiation of MigrationService to concrete ApplicationService implementations. Let MigrationService implement Service and rename runMigrations to initialize.
We want to re-use the application library in Android, but Update and Migration handling is specific for Desktop or headless apps.
