+++ apps/desktop/desktop-app/build.gradle.kts
@@ -50,6 +50,7 @@ dependencies {
     implementation("bisq:presentation")
     implementation("bisq:bisq-easy")
     implementation("bisq:application")
+    implementation("bisq:evolution")
 
     implementation(project(":desktop"))
