+++ apps/desktop/desktop/build.gradle.kts
@@ -34,6 +34,7 @@ dependencies {
     implementation("bisq:presentation")
     implementation("bisq:bisq-easy")
     implementation("bisq:application")
+    implementation("bisq:evolution")
 
     implementation("network:network-common")
     implementation("network:network")
