+++ apps/desktop/desktop/src/main/java/bisq/desktop/overlay/update/UpdaterView.java
@@ -25,8 +25,8 @@ import bisq.desktop.components.table.BisqTableView;
 import bisq.desktop.overlay.OverlayModel;
 import bisq.i18n.Res;
 import bisq.presentation.formatters.PercentageFormatter;
-import bisq.updater.DownloadItem;
-import bisq.updater.UpdaterUtils;
+import bisq.evolution.updater.DownloadItem;
+import bisq.evolution.updater.UpdaterUtils;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.SimpleBooleanProperty;
