+++ apps/seed-node-app/build.gradle.kts
@@ -11,9 +11,10 @@ dependencies {
     implementation("bisq:persistence")
     implementation("bisq:security")
     implementation("bisq:bonded-roles")
-    implementation("bisq:application")
     implementation("bisq:identity")
     implementation("bisq:user")
+    implementation("bisq:application")
+    implementation("bisq:evolution")
 
     implementation("network:network-common")
     implementation("network:network")
