+++ evolution/src/test/java/bisq/evolution/migration/MigrationServiceTests.java
@@ -1,7 +1,8 @@
-package bisq.application.migration;
+package bisq.evolution.migration;
 
 import bisq.common.platform.InvalidVersionException;
 import bisq.common.platform.Version;
+import bisq.evolution.migration.MigrationService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
