+++ settings.gradle.kts
@@ -29,6 +29,7 @@ include("chat")
 include("common")
 include("contract")
 include("daemon")
+include("evolution")
 include("identity")
 include("i18n")
 include("offer")
